requirements.txt
streamlit-nightly

pyright
pytest
